Create node and tree class and build tree method
The node class has the comparable method and an initalization that
stores the data, and left and right nodes. The tree initialization
just has the root node stored. The build tree takes an array and
a start and finish index. It then recursively calls the build tree
method until the start becomes greater than the end.