Add delete and min_value methods
The delete method finds the data to be deleted by starting at the
root and comparing the roots data to the data to be deleted. Based
on whether the data to be deleted is lower or greater than the roots
value it recursively calls the method with either the right node
or left node but keeps the data to be deleted the same. If it never
finds the data to be deleted, you'll eventually call a node that
has is nil. The method checks for this at the start and if the node
we are on is nil it returns nil. This keeps everything the same. Once
it finds the data it determines if there are any children. If there
are none it returns nil. If one it returns the one child. If two,
it goes to right and determines the lowest value and sets the data
equal to the lowest value. Then it recursively calls itself on the
right node to delete the lowest node.