Finish methods for depth and height
The height accepts a node and a count as an argument. If the left and
right nodes are nil it will return the count. Otherwise, it is
called recursively with the count increased by 1. The function
finds the maximum count. The depth method starts at root, and adds
a counter until it finds the correct node.