Add balanced? and rebalance
Balanced? checks if current node has one or two branches. If it has
two, it checks if each is balanced, and then recursively calls itself
on the left and right branch. If there is one branch, it returns false
if the height of branch node is anything but zero. Rebalance just gets
the inorder array and rebuilds the tree.